+++ packages/scroll-lock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sv/scroll-lock
 
+## 2.0.0-dev.0 (2025-10-16)
+
+### Breaking Changes
+
+- Move chart and animation elemnts into own package
+
 ## 1.1.1 (2025-04-15)
 
 ### Fixes